requirements.txt
fastapi==0.89.1
pydantic==1.10.4
uvicorn==0.20.0
SQLAlchemy==1.4.46
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==1.3.0